Add ModeSimulation classes
