+++ tidy3d/components/mode/derivatives.py
